satisfyBackfillingJobs(), backfillJobs(), processOfBackfilling()
the above functions are done.  now the left task is to get the
job_estimated_exec_time from job classad.

a new idea is we do not take the current cycle of allocations into
consideration, as busy_resources is not  strictly right as it did not
add the current allocations which are already deleted from idle......